switch default to buffered
