+++ pkg/migration/README.md
@@ -46,9 +46,9 @@ To answer this question, we need to understand that there are two types of locks
 
 So when we describe Spirit as a "non-blocking schema change tool" that is a bit of a white lie. What it means is that we don't require a metadata lock for the entire 10h schema change, as built-in MySQL DDL often does. It does not mean that we do not require locks.
 
-The common **data lock** concern is the source table. The default (unbuffered) copier issues `INSERT IGNORE INTO _new ... SELECT FROM original`, and the `SELECT` side of `INSERT ... SELECT` takes shared row locks rather than using MVCC. Production workloads touching the same rows can contend with those locks. The main knob to mitigate this is `target-chunk-time`: smaller chunks mean each copy statement holds locks for less time.
+The common **data lock** concern is the source table. By default Spirit uses the buffered copier, which reads rows into Spirit and writes them through the applier (`REPLACE INTO _new VALUES (...)`) instead of running `INSERT ... SELECT`, so it takes no locks on the source. The change source is likewise always buffered — it applies binlog row images and never runs `SELECT FROM original` — so the default path eliminates source-side data lock contention end-to-end. See [pkg/change/README.md](../change/README.md) for the subscription design.
 
-If hot rows on the source table make even short-chunk contention unacceptable, use `--buffered`. The buffered copier reads rows into Spirit and writes them through the applier instead of running `INSERT ... SELECT`, so it takes no locks on the source. The change source is already always buffered — it applies binlog row images via `REPLACE INTO _new VALUES (...)` and never touches the source — so `--buffered` plus the always-buffered subscription eliminates source-side data lock contention end-to-end. See [pkg/change/README.md](../change/README.md) for the subscription design.
+The legacy `--unbuffered` copier instead issues `INSERT IGNORE INTO _new ... SELECT FROM original`, and the `SELECT` side of `INSERT ... SELECT` takes shared row locks rather than using MVCC. Production workloads touching the same rows can contend with those locks. When using `--unbuffered`, the main knob to mitigate this is `target-chunk-time`: smaller chunks mean each copy statement holds locks for less time.
 
 The following are cases where **metadata locks** are required:
