+++ pkg/migration/check/check.go
@@ -43,7 +43,6 @@ type Resources struct {
 	Password           string
 	TLSMode            string
 	TLSCertificatePath string
-	Buffered           bool
 	// GTID, when true, opts the migration into the experimental GTID-based
 	// change source. The configuration check uses this to additionally
 	// validate gtid_mode and enforce_gtid_consistency on the source.
